Add the simplest UI of changing the Location of iPad
I have changed the Segue mode from “show” to “show detail” and added
the simplest UI of the 3rd function — change or edit the location of
iPad. I tried to use the default keyboard to test this UI controller,
but I failed since simulation did’t have the return button. So I added
a button for returning the text that users edited. I may change the
structure after this program in my iPad with its own keyboard which has
the Return button.