Fix the problem of the program can't save the location data
Every time I use segue to move from one page to another, the label
which should have stored the location which users edited didn’t show
the correct message. Then I use a globe var to keep the string and it
will keep the data as long as the program is running. But I still
wonder whether should I save the data to the local memory so even the
program is not running, it will still know the location?