Delete Error UI page and add new page for Sign Up with PickerView
I was trying to add a PickerView in my UI page but the app crashed
every time when I launched the Sign Up page. After testing in another
project, I found the problem is the page itself, so I deleted the page
and add new one. Then I add and make the PickerView work successfully.
I will finish the UI design of this page tomorrow. And I think I will
need the scroll and if it is needed, my progress may be delayed for a
little while.