Change the SignUp Page from ViewControl to TableControl
I changed the ViewControl to TableControl since it has Static Mode and
it is more suitable for the SignUp page. And I also use the Navigation
Controller to do the “Back to HomePage” job instead of a button. It
makes the page clearly. Since the TableView has Scroll Function, so I
don’t need put the table on a Scroll Sheet.